Update cre2.go to allow for compilation on Windows with MSYS2 and MingW-w64

+++ internal/cre2/cre2.go
@@ -4,6 +4,7 @@ package cre2
 
 /*
 #include <stdbool.h>
+#include <stdint.h>
 
 void* cre2_new(void* pattern, int pattern_len, void* opts);
 void cre2_delete(void* re);
@@ -25,7 +26,7 @@ void cre2_opt_set_posix_syntax(void* opt, int flag);
 void cre2_opt_set_case_sensitive(void* opt, int flag);
 void cre2_opt_set_latin1_encoding(void* opt);
 
-void* malloc(unsigned long size);
+void* malloc(size_t size);
 void free(void* ptr);
 */
 import "C"
@@ -107,7 +108,7 @@ func OptSetLatin1Encoding(opt unsafe.Pointer) {
 }
 
 func Malloc(size int) unsafe.Pointer {
-	return C.malloc(C.ulong(size))
+	return C.malloc(C.uint64_t(size))
 }
 
 func Free(ptr unsafe.Pointer) {
